Re-format link description for the first part (until TC)
Descriptions for the links should be shortened. We want to make
something very easy to read: no long sentences, just a few words. Also
remove subjective commenting, I don't believe this should be the point
in an awesome list (“a very nice introduction…”, “an excellent work
about…” etc.)

Additionally, move the two links about additional list of resources to
the end of the section (and a link to “Dive into BPF”).