Restart list with Dive into BPF List (#3)
* Restart list with Dive into BPF List

https://qmonnet.github.io/whirl-offload/2016/09/01/dive-into-bpf/\#what-is-bpf
https://raw.githubusercontent.com/qmonnet/whirl-offload/gh-pages/_posts/2016-09-01-dive-into-bpf.md

* TODO: Improved intro/about eBPF

* Drop troubleshooting stuff for now, maybe bring back some in dev section later

* Open NFP Offload Guide moved

* EBPF and Linux TC Offloada link moved

* Looks like tools/net moved

The following work:
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/tools/net?h=v4.14
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/tools/bpf

* General not generic makes more sense, imho

* Use remarkjs to format/lint markdown

* Revert "General not generic makes more sense, imho"

This reverts commit 91864365374748db0cbf4d1862ce86b59100433e.

* link to general not french slideshare

* Formatting and getting rid of italicised and underlined stuff

* Merge Documentation and Presentations into Learning Resources

* WIP: Tidying up links, and trying to find ones to drop

* Fix capitalisation for “eBPF” in title

* Move acknowledgement to bottom of file

No need to have it at the top, makes is less easy to start the reading.
As far as I'm concerned, we could even get rid of it, I don't really
mind.

* Get rid of some old and outdated links

* Re-format link description for the first part (until TC)

Descriptions for the links should be shortened. We want to make
something very easy to read: no long sentences, just a few words. Also
remove subjective commenting, I don't believe this should be the point
in an awesome list (“a very nice introduction…”, “an excellent work
about…” etc.)

Additionally, move the two links about additional list of resources to
the end of the section (and a link to “Dive into BPF”).

* Remove subsection about TC

This is about eBPF, not TC.

* Format the rest of the first section (Internals → Projects based on BPF)

* Move some stuff around

* Remove line breaks between list items

I think most of the awesome lists I've seen do not use empty lines
between each list item, but I'm not even sure I remember correctly about
that. I guess in the end it's only a matter of personal preferences.

* Get rid of some italics

* Hardware offload tidy up, I think

* Make eBPF-based project a main section i.o. a subsection

Enough contents to make it a section on its own. And first section is
too large anyway.

* Rework section on BPF development

* Create section on tools and rework its contents

* Move subsection on other lists to section at the end of the document

Did not really feel at its place in the middle of the document.

* Note about people providing feedback