Reorganise the list: hightlight documentation, raise tooling
This commit brings some reorganisation to the list. The “reference”
documentation, which encompasses kernel documentation, manual pages, and
the exhaustive Cilium guide, are moved to the top of the list to
better show that they are the main pieces of eBPF documentation
containing and should ideally be consulted first when searching for
technical information.

The rest of the documentation resources is in a separate section for
“articles and presentations”.

Because it is essential to indicate to new eBPF users what are the tools
to use to work with eBPF, raise the “tools and utilites” section in the
list, before the eBPF-related projects. The links to code section can
come later.

Additionally, add title case to some titles to have everything
consistent. And update the table of contents to reflect the changes.

No link has been added or removed from the list.