greggd: add a period at the end of the description
I don't remember how to run the linter, but I'm pretty sure that it
is supposed to complain about missing periods at the end of
descriptions. We should probably look into running it in a GitHub action
in the future.

Anyway. Add a period at the end of greggd's description.